# backend/docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
  models.ProxyDownloadRequest:
    properties:
      video_url:
        type: string
    required:
    - video_url
    type: object
  models.QualitiesResponse:
    properties:
      available_qualities:
        items:
          $ref: '#/definitions/models.QualityOption'
        type: array
      platform:
        type: string
    type: object
  models.QualityOption:
    properties:
      height:
        type: integer
      label:
        type: string
      quality:
        type: string
      video_url:
        type: string
      width:
        type: integer
    type: object
  models.QualityRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  models.VideoRequest:
    properties:
      quality:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  models.VideoResponse:
    properties:
      available_qualities:
        items:
          $ref: '#/definitions/models.QualityOption'
        type: array
      duration:
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      platform:
        type: string
      processed_at:
        type: string
      quality:
        type: string
      title:
        type: string
      video_url:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@vidtogallery.com
    name: VidToGallery API Support
  description: API for downloading videos from social media platforms (Instagram,
    Twitter/X, YouTube)
  title: VidToGallery API
  version: 1.0.0
paths:
  /api/v1/download:
    post:
      consumes:
      - application/json
      description: Download video from social media platform with specified quality
      parameters:
      - description: Video URL and quality to download
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Video downloaded successfully
          schema:
            $ref: '#/definitions/models.VideoResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Download video with quality
      tags:
      - Video Processing
  /api/v1/proxy-download:
    post:
      consumes:
      - application/json
      description: Download video file through backend proxy to avoid CORS restrictions
      parameters:
      - description: Video URL to proxy download
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProxyDownloadRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Video file
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Proxy download video file
      tags:
      - Video Processing
  /api/v1/qualities:
    post:
      consumes:
      - application/json
      description: Get list of available video qualities for a social media URL
      parameters:
      - description: Video URL to check qualities for
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.QualityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Available qualities retrieved successfully
          schema:
            $ref: '#/definitions/models.QualitiesResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get available video qualities
      tags:
      - Video Processing
  /health:
    get:
      description: Check if the API is running and healthy
      produces:
      - application/json
      responses:
        "200":
          description: API is healthy
          schema:
            additionalProperties: true
            type: object
      summary: Health check endpoint
      tags:
      - Health
schemes:
- http
- https
swagger: "2.0"